Refactor model and custom GUI files into subpackages for neatness
CustomGUI for custom or modified qtgui elements like CodeEditor.
ModelEditors and its subpackages for model editors
ModelEditors.Dialogs for dialogs called by model editors
